Move site name column to a separate component.
